✨ validate create event 🗑 refactor add friend to event
