+++ src/components/CreateEventComponents/CreateEvent.js
@@ -3,22 +3,25 @@ import {
   View,
   Text,
   StyleSheet,
-  TextInput,
+  ScrollView,
   TouchableOpacity,
   TouchableWithoutFeedback,
   Keyboard,
 } from "react-native";
+import { Button, Input, Icon, Overlay } from "react-native-elements";
 import firebase from "firebase";
 import DateTimePicker from "@react-native-community/datetimepicker";
+import InviteFriendCard from "./InviteFriendCard";
 
 const CreateEvent = ({ route, navigation }) => {
   // --- State ----------------- //
+  const [user, setUser] = useState(null);
   const [location, setLocation] = useState("");
   const [datetime, setDateTime] = useState(new Date());
   const [description, setDescription] = useState("");
   const [friends, setFriends] = useState([]);
-  const [length, setLength] = useState(friends.length);
   const [loading, setLoading] = useState(false);
+  const [open, setOpen] = useState(false);
   // --------------------------- //
 
   // --- Helpers --------------- //
@@ -34,9 +37,24 @@ const CreateEvent = ({ route, navigation }) => {
     setLoading(false);
   };
 
-  const addFriend = (length, friendArr) => {
-    setLength(length);
-    setFriends(friendArr);
+  function arrayRemove(arr, value) {
+    return arr.filter(function (ele) {
+      return ele != value;
+    });
+  }
+
+  const addSelected = (selected, friendName) => {
+    if (!selected) {
+      var tempArr = friends;
+      tempArr.push(friendName);
+      setFriends(tempArr);
+    }
+
+    if (selected && friends.includes(friendName)) {
+      var tempArr = friends;
+      tempArr = arrayRemove(tempArr, friendName);
+      setFriends(tempArr);
+    }
   };
   // --------------------------- //
 
@@ -69,25 +87,72 @@ const CreateEvent = ({ route, navigation }) => {
   };
   // --------------------------- //
 
+  // --- Read DB --------------- //
+  useEffect(() => {
+    const unsubscribe = firebase
+      .firestore()
+      .collection("users")
+      .where("email", "==", route.params.user)
+      .onSnapshot((snapshot) => {
+        setUser(
+          snapshot.docs.map((doc) => ({
+            id: doc.id,
+            u: doc.data(),
+          }))
+        );
+      });
+    return () => {
+      unsubscribe();
+    };
+  }, []);
+  // --------------------------- //
+
   return (
     <TouchableWithoutFeedback onPress={Keyboard.dismiss} accessible={false}>
       <View style={styles.container}>
         <Text style={styles.title}>
           Hi {route.params.user}, let's create an event!
         </Text>
+        <Overlay
+          isVisible={open}
+          onBackdropPress={() => setOpen(!open)}
+          fullScreen
+        >
+          <TouchableOpacity style={styles.exit} onPress={() => setOpen(false)}>
+            <Icon name="cancel" size={25} color="black" />
+          </TouchableOpacity>
+          <ScrollView>
+            {user ? (
+              user.map(({ id, u }) =>
+                u.friends.map((friend) => (
+                  <InviteFriendCard
+                    key={id + friend}
+                    friend={friend}
+                    invitedFriends={friends}
+                    addSelected={addSelected}
+                  />
+                ))
+              )
+            ) : (
+              <React.Fragment />
+            )}
+          </ScrollView>
+        </Overlay>
 
-        <TextInput
+        <Input
           style={styles.input}
           onChangeText={setDescription}
           placeholder="Name or description"
           value={description}
+          errorMessage={validate("description", description)}
         />
 
-        <TextInput
+        <Input
           style={styles.input}
           onChangeText={setLocation}
           placeholder="Location"
           value={location}
+          errorMessage={validate("location", location)}
         />
 
         <View style={styles.datePickers}>
@@ -103,13 +168,7 @@ const CreateEvent = ({ route, navigation }) => {
 
         <TouchableOpacity
           style={styles.inviteFriends}
-          onPress={() =>
-            navigation.push("InviteFriends", {
-              email: route.params.user,
-              friends: friends,
-              paramAddFriend: addFriend,
-            })
-          }
+          onPress={() => setOpen(true)}
         >
           <Text>+Friends {friends.length}</Text>
         </TouchableOpacity>
@@ -133,6 +192,23 @@ const CreateEvent = ({ route, navigation }) => {
   );
 };
 
+const validate = (name, value) => {
+  switch (name) {
+    case "description":
+      if (value === "") {
+        return "A description is required";
+      }
+      break;
+    case "location":
+      if (value === "") {
+        return "A location is required";
+      }
+      break;
+  }
+
+  return "";
+};
+
 const styles = StyleSheet.create({
   scrollView: {
     backgroundColor: "#fff",
@@ -145,7 +221,7 @@ const styles = StyleSheet.create({
     padding: 15,
   },
   container: {
-    marginTop: "15%",
+    marginTop: "10%",
     flex: 1,
     height: "100%",
   },
@@ -182,6 +258,10 @@ const styles = StyleSheet.create({
     alignSelf: "center",
     marginTop: 10,
     width: "60%",
+  },
+  exit: {
+    alignSelf: "flex-end",
+    padding: 5,
   },
 });
